# Support recursive calls that respect merged functions