--- a/typed-function.js
+++ b/typed-function.js
@@ -750,12 +750,8 @@
      * @param {function} fn
      * @return {function} Returns a wrapped function
      */
-    function compileArgsPreprocessing(params, fn, resolveSelf) {
-      function fnResolveSelf() {
-        return fn.apply(resolveSelf(), arguments);
-      }
-
-      var fnConvert = fnResolveSelf;
+    function compileArgsPreprocessing(params, fn) {
+      var fnConvert = fn;
 
       // TODO: can we make this wrapper function smarter/simpler?
 
@@ -773,7 +769,7 @@
             args[last] = arguments[last].map(compiledConversions[last]);
           }
 
-          return fnResolveSelf.apply(null, args);
+          return fn.apply(this, args);
         }
       }
 
@@ -782,7 +778,7 @@
         var offset = params.length - 1;
 
         fnPreprocess = function preprocessRestParams () {
-          return fnConvert.apply(null,
+          return fnConvert.apply(this,
               slice(arguments, 0, offset).concat([slice(arguments, offset)]));
         }
       }
@@ -1047,15 +1043,9 @@
       var test41 = ok4 ? compileTest(signatures[4].params[1]) : notOk;
       var test51 = ok5 ? compileTest(signatures[5].params[1]) : notOk;
 
-      var fn;
-
-      function resolveSelf() {
-        return fn;
-      }
-
       // compile the functions
       var fns = signatures.map(function(signature) {
-        return compileArgsPreprocessing(signature.params, signature.fn, resolveSelf);
+        return compileArgsPreprocessing(signature.params, signature.fn);
       });
 
       var fn0 = ok0 ? fns[0] : undef;
@@ -1080,7 +1070,7 @@
 
         for (var i = iStart; i < iEnd; i++) {
           if (tests[i](arguments)) {
-            return fns[i].apply(null, arguments);
+            return fns[i].apply(this, arguments);
           }
         }
 
@@ -1089,18 +1079,24 @@
 
       // create the typed function
       // fast, specialized version. Falls back to the slower, generic one if needed
-      fn = function fn(arg0, arg1) {
-        'use strict';
+      var makeFn = function () {
+        var fn = function fn(arg0, arg1) {
+          'use strict';
 
-        if (arguments.length === len0 && test00(arg0) && test01(arg1)) { return fn0.apply(null, arguments); }
-        if (arguments.length === len1 && test10(arg0) && test11(arg1)) { return fn1.apply(null, arguments); }
-        if (arguments.length === len2 && test20(arg0) && test21(arg1)) { return fn2.apply(null, arguments); }
-        if (arguments.length === len3 && test30(arg0) && test31(arg1)) { return fn3.apply(null, arguments); }
-        if (arguments.length === len4 && test40(arg0) && test41(arg1)) { return fn4.apply(null, arguments); }
-        if (arguments.length === len5 && test50(arg0) && test51(arg1)) { return fn5.apply(null, arguments); }
+          if (arguments.length === len0 && test00(arg0) && test01(arg1)) { return fn0.apply(fn, arguments); }
+          if (arguments.length === len1 && test10(arg0) && test11(arg1)) { return fn1.apply(fn, arguments); }
+          if (arguments.length === len2 && test20(arg0) && test21(arg1)) { return fn2.apply(fn, arguments); }
+          if (arguments.length === len3 && test30(arg0) && test31(arg1)) { return fn3.apply(fn, arguments); }
+          if (arguments.length === len4 && test40(arg0) && test41(arg1)) { return fn4.apply(fn, arguments); }
+          if (arguments.length === len5 && test50(arg0) && test51(arg1)) { return fn5.apply(fn, arguments); }
 
-        return generic.apply(null, arguments);
+          return generic.apply(fn, arguments);
+        }
+
+        return fn;
       }
+
+      var fn = makeFn();
 
       // attach name the typed function
       try {