This PR implements a method of recursion that works with merged typed-functions.  The issue was originally discussed here: https://github.com/josdejong/mathjs/issues/1885.

I tested this change against math.js, replacing all recursive typed-function calls I could find in the library, and the tests are passing.  I can open a PR for that too, if this looks acceptable.

A slight modification on this approach would be to remove the proxy object and change 

```
return fn.apply(proxy, arguments);
```

to

```
return fn.apply(self, arguments);
```

which would then allow definitions like so:

```
var sqrt = typed({
  'number': function (value) {
    return Math.sqrt(value);
  },
  'string': function (value) {
    return this(parseInt(value, 10));
  }
});
```

However, I am not sure if using a function bound as `this` works in all browsers.