@josdejong 2020-06-27
This is a really nice solution Nick :sunglasses:

It's a very smart good idea to utilize the function's `this`, which we can determine ourselves when calling `fn.apply`. I'm not sure either if this will give issues in certain browsers, but I don't expect that. We should definitely give that a try. It would be a great fit since it works very natural and doesn't need a new API at all :).

If I'm not mistaken, when using the function's `this` (without the `proxy `and `baseFn`), it will be as simple as the following, right?

```js
function compileArgsPreprocessing(params, fn, resolveSelf) {
  // ...
  return fn.apply(resolveSelf(), args);
  // ...
}
```

We should also double check if this has any impact on the performance.