@nickewing 2020-06-28
Thanks for the feedback.  I think it's a good call to give directly binding the function a try.

Given the way `compileArgsPreprocessing` conditionally wraps the `fn` function, we can't completely remove the wrapper function (previously `baseFn`).

I refactored `compileArgsPreprocessing` since there was already a note in there that it could use some cleanup.  If you think the refactor is too much, I can revert that commit.