@josdejong 2020-06-29
That is a really neat refactoring, thanks 👍 

I tested in mathjs and it indeed just works replacing self references with `this`. I did run all tests with this on all browsers including IE, and it works like a charm.

I did run the benchmark in `benchmark/benchmark.js`, and this shows a serious decrease in performance. I think that needs some attention:

```
// node benchmark/benchmark.js

before recursive support:

typed add            x 32,011,095 ops/sec ±0.42% (89 runs sampled)
native add           x 34,924,831 ops/sec ±0.32% (94 runs sampled)

after recursive support:

typed add            x 23,669,110 ops/sec ±0.71% (95 runs sampled)
native add           x 34,913,476 ops/sec ±0.48% (95 runs sampled)
```

Can this be related to the extra indirection introduce with `applyWithSelf`? It may be worth trying to revert the refactorings, start from there adding `resolveSelf()` to see if this is a consequence of `resolveSelf()` or of the refactoring. And we should also be careful that maybe this micro benchmark can be off too. What do you think?