@nickewing 2020-07-01
Good call on checking the benchmarks.  It seems that it wasn't actually the refactored `compileArgsPreprocessing` function that caused the slowness because after reverting that, I was still getting benchmark numbers similar to what you saw.  It must have been something to do with using closure in the `resolveSelf` that was passed into `compileArgsPreprocessing`.

I reworked things again and now the numbers look to be comparable.  There are couple slower runs in this branch and in the `develop` branch, but given that the native runs were also slow there, I'm guessing those aren't relevant.  I did leave out the other function refactoring though given the changes to call `resolveSelf` were no longer necessary.

I also ran the new implementation against math.js, with all the recursive function calls replaced with `this`, and things looked ok there too.

recursion branch                                                    | develop branch
--------------------------------------------------------------------|--------------------------------------------------------------------
typed add            x 31,814,191 ops/sec ±0.23% (97 runs sampled)  | typed add            x 30,844,791 ops/sec ±2.01% (90 runs sampled)
native add           x 36,088,603 ops/sec ±0.51% (94 runs sampled)  | native add           x 36,104,421 ops/sec ±0.23% (99 runs sampled)
typed add            x 23,485,833 ops/sec ±0.24% (96 runs sampled)  | typed add            x 30,438,337 ops/sec ±0.28% (87 runs sampled)
native add           x 25,594,284 ops/sec ±0.28% (92 runs sampled)  | native add           x 35,077,959 ops/sec ±0.25% (99 runs sampled)
typed add            x 31,598,207 ops/sec ±0.23% (94 runs sampled)  | typed add            x 30,879,719 ops/sec ±0.24% (92 runs sampled)
native add           x 35,506,001 ops/sec ±0.21% (94 runs sampled)  | native add           x 35,375,359 ops/sec ±0.22% (97 runs sampled)
typed add            x 30,558,612 ops/sec ±0.28% (91 runs sampled)  | typed add            x 31,482,917 ops/sec ±0.22% (93 runs sampled)
native add           x 34,990,322 ops/sec ±0.32% (91 runs sampled)  | native add           x 35,711,849 ops/sec ±0.18% (98 runs sampled)
typed add            x 25,565,964 ops/sec ±1.51% (96 runs sampled)  | typed add            x 31,196,506 ops/sec ±0.31% (92 runs sampled)
native add           x 28,363,794 ops/sec ±0.10% (96 runs sampled)  | native add           x 36,093,807 ops/sec ±0.17% (96 runs sampled)
typed add            x 31,570,449 ops/sec ±0.25% (94 runs sampled)  | typed add            x 25,673,182 ops/sec ±0.28% (93 runs sampled)
native add           x 35,816,558 ops/sec ±1.36% (97 runs sampled)  | native add           x 28,387,732 ops/sec ±0.14% (96 runs sampled)
typed add            x 31,125,150 ops/sec ±0.43% (94 runs sampled)  | typed add            x 30,437,918 ops/sec ±0.28% (96 runs sampled)
native add           x 35,639,589 ops/sec ±0.17% (97 runs sampled)  | native add           x 34,852,741 ops/sec ±0.21% (94 runs sampled)
typed add            x 31,335,404 ops/sec ±0.51% (92 runs sampled)  | typed add            x 31,240,131 ops/sec ±0.44% (91 runs sampled)
native add           x 35,098,381 ops/sec ±0.20% (95 runs sampled)  | native add           x 36,110,116 ops/sec ±0.15% (98 runs sampled)
typed add            x 31,178,082 ops/sec ±0.30% (94 runs sampled)  | typed add            x 31,251,682 ops/sec ±0.37% (92 runs sampled)
native add           x 35,338,348 ops/sec ±0.16% (94 runs sampled)  | native add           x 36,057,924 ops/sec ±0.25% (95 runs sampled)
typed add            x 25,724,821 ops/sec ±0.28% (94 runs sampled)  | typed add            x 31,288,245 ops/sec ±0.25% (90 runs sampled)
native add           x 28,578,597 ops/sec ±0.15% (97 runs sampled)  | native add           x 36,064,112 ops/sec ±0.20% (94 runs sampled)